Fix iOS simulator release asan build after 268314@main https://bugs.webkit.org/show_bug.cgi?id=261977 rdar://115910638
Reviewed by Simon Fraser.

Something about my attempt to put RefCounted and vtable overhead in the same
base class wasn't quite the same as having them separate.  This switches to
actually use RefCounted when trying to make a struct that is the same size as
something that is RefCounted for the purposes of verifying that all members
are serialized.

* Source/WebCore/platform/network/NetworkLoadMetrics.h:
* Source/WebKit/Scripts/generate-serializers.py:
(check_type_members.is):
(generate_impl):
* Source/WebKit/Scripts/webkit/tests/GeneratedSerializers.cpp:
(VirtualTableOverhead<true>::~VirtualTableOverhead):
* Source/WebKit/Scripts/webkit/tests/WebKitPlatformGeneratedSerializers.cpp:
(VirtualTableOverhead<true>::~VirtualTableOverhead):

Canonical link: https://commits.webkit.org/268360@main