[WGSL] GlobalVariableRewriter doesn't distinguish between multiple globals with the same binding https://bugs.webkit.org/show_bug.cgi?id=262045 rdar://115994053
Reviewed by Dan Glastonbury.

When GlobalVariableRewriter determines which globals are used by each entrypoint, it
uses a pair of (group, binding) to determine which globals are used. However, it is
valid to have multiple globals with the same (group, binding) pair, so long as they are
not used by the same entrypoint. In those cases, we ended up marking all such variables
as used, which resulted in invalid code since we had multiple struct fields with the same
[[id]]. The fix is trivial, since we already keep track of the names of all the variables
used, all we need to do is add an extra check to see if the variable is used according
to its name.

* Source/WebGPU/WGSL/GlobalVariableRewriter.cpp:
(WGSL::RewriteGlobalVariables::insertStructs):
* Source/WebGPU/WGSL/tests/valid/global-same-binding.wgsl: Added.

Canonical link: https://commits.webkit.org/268435@main