[WGSL] Extra trailing semicolons cause parsing failure https://bugs.webkit.org/show_bug.cgi?id=262044 rdar://115993754
Reviewed by Dan Glastonbury.

Although we skip extraneous semicolons at the begining of parseGlobalDecl and
parseStatement, that only works if there is declaration/statement after the
skipped semicolon. If it is a trailing semicolon, these functions will fail
as they expect to see the start of a declaration/statement after that.
In order to fix it, we move the skipping into the functions that consume
a list of declarations/statements, as they already detect whether they are
at the end of the list.

* Source/WebGPU/WGSL/Parser.cpp:
(WGSL::Parser<Lexer>::parseShader):
(WGSL::Parser<Lexer>::parseGlobalDecl):
(WGSL::Parser<Lexer>::parseStatement):
(WGSL::Parser<Lexer>::parseCompoundStatement):
* Tools/TestWebKitAPI/Tests/WGSL/ParserTests.cpp:
(TEST):

Canonical link: https://commits.webkit.org/268437@main