WebRTC maintain-framerate degradationPreference does not apply to resolution adaptation https://bugs.webkit.org/show_bug.cgi?id=261266 rdar://115481961
Reviewed by Jean-Yves Avenard.

CPU adaptation needs to be turned on to allow degradationPreference, this gets us closer to how Chrome is setting up PC configurations..
We initially turned off CPU adaptation as it was creating some issues on bots.
If it turns out to still be the case, we will migrate tests to use maintain-resolution.

* Source/WebCore/Modules/mediastream/libwebrtc/LibWebRTCPeerConnectionBackend.cpp:
(WebCore::configurationFromMediaEndpointConfiguration):

Canonical link: https://commits.webkit.org/268443@main