[GTK][WPE] Gardening LayoutTests/webrtc/video-disabled-black.html
Unreviewed test gardening.

Test was gardened a couple of weeks ago in webkit.org/b/261024; and
webkit.org/b/254212 is also related to this test as a fix. Since
webkit.org/b/254212, test has been flaky; prior to that, it failed in bots.
Generally a flaky test as noted in 261024 which is still open.

Noted in existing bugs 261024, 254212; moved expectations up to glib.

* LayoutTests/platform/glib/TestExpectations: move expectation here for flakiness
* LayoutTests/platform/gtk/TestExpectations:

Canonical link: https://commits.webkit.org/268498@main