[css-grid] adjustGridPositionsFromStyle should set a subgrid's indefinite span to use the null string https://bugs.webkit.org/show_bug.cgi?id=260527 rdar://114271485
Reviewed by Matt Woodrow.

If a subgrid has an indefinite span in the subgridded dimension,
adjustGridPositionsFromStyle will attempt to set the span size based
on the the number of named lines that it has. This is done with a call
similar to span.setSpanPosition(std::max(1, lineCount - 1), emtpyString()).

This code presumably does not want to set the m_namedGridLine portion
of the GridPosition as we are not actually referencing a named grid
line. However, we should probably be using String() here instead of
emptyString() to set that value to the null string as the rest of the
code checks against !position.namedGridLine().isNull() when attempting
to use the named grid line. This change helps us avoid entering code
guarded by that type of logic when we probably do not want to.

* LayoutTests/TestExpectations:
* Source/WebCore/rendering/style/GridPositionsResolver.cpp:
(WebCore::adjustGridPositionsFromStyle):

Canonical link: https://commits.webkit.org/268510@main