[visionOS] Sign in with Apple window not anchored to the host app https://bugs.webkit.org/show_bug.cgi?id=262230 rdar://116142970
Reviewed by NOBODY (OOPS!).

When the Sign in with Apple window is presented, AuthKit looks for a scene ID
to anchor the window to first, and uses the bundle identifier as a fallback to
grab an appropriate window to anchor to.

When an app uses an ASWebAuthenticationSession to handle the login flow, it uses
SVS to present an SFAuthenticationViewController. In this case, AuthKit
receives the wrong bundle identifier (com.apple.SafariViewService instead of the
host app), and we are not currently providing a scene identifier to AuthKit, so
it cannot find a reasonable window to anchor the sign in sheet to. This leads to
the issue described in the radar, where the SiwA sheet is presented in
unexpected places from a bincompat app using web authentication

AuthKit has made changes to accept a callerSceneIndetifier from the auth request
(ASAuthorizationProviderExtensionAuthorizationRequest). So the fix is to return
the correct sceneID in this class.

WebKit (in SOAuthorizationSession) is responsible for firing off the request,
which is actually an SOAuthorizationRequest. Later, AuthKit wraps this request
in the AS..AuthorizationRequest. So, to avoid modifying SOAuthorizationRequest
(which lives in AppSSO code), we hijack the authorizationOptions dictionary to
stuff in the `callerSceneIdentifier`, which we can return for the newly added
property in AS...AuthorizationRequest.

WebKit calls out to its UIDelegate to grab the correct scene identifier from
SVS, and sets it on the authorizationOptions dictionary of the request.

* Source/WebKit/UIProcess/API/Cocoa/WKUIDelegatePrivate.h:
Add a delegate method to retrieve the callerSceneIdentifier on visionOS.
This method is implemented in SafariViewService.

* Source/WebKit/UIProcess/Cocoa/SOAuthorization/SOAuthorizationSession.mm:
(WebKit::SOAuthorizationSession::continueStartAfterDecidePolicy):
Set the callerSceneIdentifier on the SOAuthorizationRequest options dict
to allow AuthenticationServices to retrieve it later and forward it to
AuthKit.