[WGSL] Add more logging to constant rewriter and clear evaluation result https://bugs.webkit.org/show_bug.cgi?id=262305 rdar://116179924
Reviewed by Mike Wyrzykowski.

There was a bug in the constant rewriter where the wrong value would be returned
for expressions. This is caused by the fact that the rewriter implementation is
rather sparse at the moment, and we weren't cleaning the previous evaluation result,
so a stale value was being returned. In order to identify the root cause of the
issue more logging was required, and it felt appropriate to leave the logging in
(behind the same flag already in use) for future debugging.

* Source/WebGPU/WGSL/ConstantRewriter.cpp:
(WGSL::ConstantRewriter::visit):
(WGSL::ConstantRewriter::evaluate):
(WGSL::ConstantRewriter::evaluated):
(WGSL::ConstantRewriter::introduceVariable):

Canonical link: https://commits.webkit.org/268637@main