text-spacing: integrate with low level iterator, fast-path: width-iterator https://bugs.webkit.org/show_bug.cgi?id=262317 rdar://115984437
Reviewed by NOBODY (OOPS!).

This is a preparation patch for text-spacing: 'auto'.
We need to design a state machine able to process text-spacing (TextSpacingEngine),
and integrate it with our WidthIterator (low-level) so we can process text-spacing
within a single run. This will be only preparation, will not actually compute
spacing or handle behavior between runs (high level).

TextRun is used to transport TextSpacingEngine's state between runs
and from high-level callers to low-level executer (WidthIterator), although
this is not yet done.

The way we should, in the future, caculate spacing for text-spacing 'auto' is
by processing pair of adjacent characters, current char and next char.
The spacing to be added is always the spacing between these two characters, divided in 2 components:
1. right-width: spacing to be added to the right of the current char
2. left-width: spacing to be added to the left of the next char

For that reason, we need to save the left-width into our engine's state
so we can pass that information later on to a next text run, when the
next char is actually not coming from the current run we are processing but
from the next run. For such a caseWe also need the next char to be saved into the state,
so it can be passed to WidthIterator while it is still processing the current run.

At each iteration, we will then add right-width to the right of the glyph
corresponding to the current char. We will add the state's left-width to the
left of the current char (since it was calculated in the last iteration) and
we will update the state's left-width with the left-width we calculate at
the current iteration.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/CSSProperties.json:
* Source/WebCore/layout/formattingContexts/inline/text/TextUtil.cpp:
(WebCore::Layout::TextUtil::canUseSimplifiedTextMeasuring):
* Source/WebCore/platform/graphics/GlyphBuffer.h:
(WebCore::GlyphBuffer::expandAdvanceToLeft):
* Source/WebCore/platform/graphics/TextRun.cpp:
* Source/WebCore/platform/graphics/TextRun.h:
(WebCore::TextRun::TextRun):
* Source/WebCore/platform/graphics/WidthIterator.cpp:
(WebCore::WidthIterator::WidthIterator):
(WebCore::WidthIterator::hasTextSpacingAuto const):
(WebCore::WidthIterator::applyExtraSpacingAfterShaping):
(WebCore::WidthIterator::advance):
* Source/WebCore/platform/graphics/WidthIterator.h:
* Source/WebCore/platform/text/TextSpacing.cpp: Added.
(WebCore::TextSpacingEngine::processTextSpacingAndUpdateState):
(WebCore::TextSpacingEngine::adjustTextSpacing):
(WebCore::calculateCJKAdjustingSpacing):
* Source/WebCore/platform/text/TextSpacing.h:
(WebCore::TextSpacingEngine::TextSpacingEngine):
* Source/WebCore/platform/text/mac/CoreTextChineseCompositionEngine.cpp:
(WebCore::ChineseCompositionRules::characterSpacing):
(characterToCharacterClass): Deleted.
(isDigit): Deleted.
(isNarrowFullWidth): Deleted.
(ChineseCompositionRules::characterClass): Deleted.
(ChineseCompositionRules::characterSpacing): Deleted.
* Source/WebCore/platform/text/mac/CoreTextChineseCompositionEngine.h:
* Source/WebCore/platform/text/mac/CoreTextCompositionEngine.cpp:
(WebCore::characterSpacing):
(WebCore::calculateCJKAdjustingSpacing):
(CompositionRules::characterSpacing): Deleted.
* Source/WebCore/platform/text/mac/CoreTextCompositionEngine.h: